rename state button to instruction button
