+++ app/assets/stylesheets/buttons.scss
@@ -140,7 +140,7 @@
   height: 25px;
 }
 
-.state-button {
+.instruction-button {
   display: inline-block;
 }
