+++ app/javascript/new_design/instruction-button.js
